Crate solana_keypair

Concrete implementation of a Solana Signer from raw bytes

Modules

seed_derivable

Implementation of the SeedDerivable trait for Keypair

signable

Structs

Address

The address of a Solana account.

Keypair

A vanilla Ed25519 key pair

Signature

SignatureError

Signature errors.

Constants

KEYPAIR_LENGTH

Traits

EncodableKey

The EncodableKey trait defines the interface by which cryptographic keys/keypairs are read, written, and derived from sources.

EncodableKeypair

The EncodableKeypair trait extends EncodableKey for asymmetric keypairs, i.e. have associated public keys.

Signer

The Signer trait declares operations that all digital signature providers must support. It is the primary interface by which signers are specified in Transaction signing interfaces

Functions

keypair_from_seed

Constructs a Keypair from caller-provided seed entropy

keypair_from_seed_phrase_and_passphrase

read_keypair

Reads a JSON-encoded Keypair from a Reader implementor

read_keypair_file

Reads a Keypair from a file

write_keypair

Writes a Keypair to a Write implementor with JSON-encoding

write_keypair_file

Writes a Keypair to a file with JSON-encoding